feat: use server actions
