+++ components/hCaptcha.tsx
@@ -1,3 +1,4 @@
+"use client";
 import HCaptcha from "@hcaptcha/react-hcaptcha";
 
 export default function Captcha() {
